+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Next
+- Fix "electron.app.hide is not a function" error when closing app window on Windows/Linux
+
 ## 1.1.0 - 2026 Aug 8
 - Add "Stop" menu item
 - Fix volume shortcuts