+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Next
+- Remove macOS 10.15 support
+- Update many dependencies
 - Fix "electron.app.hide is not a function" error when closing app window on Windows/Linux
 
 ## 1.1.0 - 2026 Aug 8