+++ CHANGELOG.md
@@ -8,8 +8,9 @@
 - Add `artist:` filtering for titles without brackets, like `Supercell feat. Veela`
 - Add `artist:` filtering for title prefixes `w/`, `prod`, `prod by` and suffixes `cover`, `rework`, `refix`, `vip`, `mashup`
 - Remove the active filter highlight
-- Fix undo/redo in the filter textbox only changing one character at a time
 - Fix "electron.app.hide is not a function" error when closing app window on Windows/Linux
+- Fix undo/redo in the filter textbox only changing one character at a time
+- Fix multi-word field filter not working for ending words, like `title:"remix"`
 
 ## 1.1.0 - 2026 Aug 8
 - Add "Stop" menu item