+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 ## Next
 - Remove macOS 10.15 support
-- Update many dependencies
-- Fix undo/redo in the filter textbox only changing one character at a time
+- Add exact filtering with double quotes, like `"the one"`
+- Show filter syntax guide on empty page
+- Add syntax highlighting for filters like `album:"mirai sekai"`
+- Add `artist:` filtering for titles without brackets, like `Supercell feat. Veela`
+- Add `artist:` filtering for title prefixes `w/`, `prod`, `prod by` and suffixes `cover`, `rework`, `refix`, `vip`, `mashup`
+- Remove the active filter highlight
 - Fix "electron.app.hide is not a function" error when closing app window on Windows/Linux
+- Fix undo/redo in the filter textbox only changing one character at a time
+- Fix multi-word field filter not working for ending words, like `title:"remix"`
 
 ## 1.1.0 - 2026 Aug 8
 - Add "Stop" menu item