+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Next
+## 1.2.0 - 2026 Aug 31
 - Remove macOS 10.15 support
 - Add exact filtering with double quotes, like `"the one"`
 - Show filter syntax guide on empty page